Подготовка dto структур и конвертеров.
