+++ src/internal/presentation/cli/state/game.go
@@ -52,27 +52,8 @@ func (g *Game) updateGameView() {
 
 	player := g.level.Player
 	if player != nil {
-		g.view.UpdatePlayerInfo(&viewcli.PlayerInfo{
-			Health:           player.Attributes.Health,
-			MaxHealth:        player.Attributes.MaxHealth,
-			Strength:         player.Attributes.Strength,
-			Agility:          player.Attributes.Agility,
-			TemporaryEffects: g.convertEffectsToView(player.TemporaryEffects),
-		})
+		g.view.UpdatePlayerInfo(dto.ConvertPlayerToDto(g.level.Player))
 	}
-}
-
-func (g *Game) convertEffectsToView(effects []*primitives.Effect) []viewcli.EffectInfo {
-	viewEffects := make([]viewcli.EffectInfo, 0, len(effects))
-	for _, effect := range effects {
-		viewEffects = append(viewEffects, viewcli.EffectInfo{
-			DurationSteps:  int(effect.Duration.Steps),
-			HealthModify:   effect.Attributes.Health,
-			StrengthModify: effect.Attributes.Strength,
-			AgilityModify:  effect.Attributes.Agility,
-		})
-	}
-	return viewEffects
 }
 
 func (g *Game) initInput() {
@@ -194,9 +175,7 @@ func (g *Game) eventToAction(event *tcell.EventKey) action.Type {
 }
 
 func (g *Game) updateItemInfo(pos primitives.Point2D[int]) {
-	item := g.level.GetItemAtPosition(pos)
-	itemInfo := dto.GetItemInfo(item)
-	g.view.UpdateItemInfo(itemInfo)
+	g.view.UpdateItemInfo(dto.ConvertPositionalItemToDto(g.level.GetItemAtPosition(pos)))
 }
 
 func (g *Game) Update(float64) signals.Type {
