+++ src/internal/presentation/dto/player_info.go
@@ -1,0 +1,25 @@
+package dto
+
+import "gogue/internal/model/entities"
+
+type PlayerInfo struct {
+	Health           float64
+	MaxHealth        float64
+	Strength         float64
+	Agility          float64
+	TemporaryEffects []EffectInfo
+}
+
+func ConvertPlayerToDto(player *entities.Player) *PlayerInfo {
+	if player == nil {
+		return nil
+	}
+
+	return &PlayerInfo{
+		Health:           player.Attributes.Health,
+		MaxHealth:        player.Attributes.MaxHealth,
+		Strength:         player.Attributes.Strength,
+		Agility:          player.Attributes.Agility,
+		TemporaryEffects: ConvertEffectsToDto(player.TemporaryEffects),
+	}
+}
